// standard_reports/item-usage/latest/convert_data_js/src/generated-types/arguments.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type Arguments = StockFilters;

export interface StockFilters {
  /**
   * Item Code
   */
  itemCode?: string;
  /**
   * Item Name
   */
  itemName?: string;
  /**
   * Average Monthly Consumption Look Back Period
   */
  monthlyConsumptionLookBackPeriod?: number;
  /**
   * sort by
   */
  sort?:
    | "name"
    | "code"
    | "SOH"
    | "MOS"
    | "monthConsumption"
    | "lastMonthConsumption"
    | "twoMonthsAgoConsumption"
    | "expiringInSixMonths"
    | "expiringInTwelveMonths"
    | "stockOnOrder"
    | "AMC12"
    | "AMC24";
  /**
   * sort by dir
   */
  dir?: "asc" | "desc";
  [k: string]: unknown;
}